requirements.txt
autopep8==1.3.5
configparser==3.5.0
Django==1.10
djangorestframework==3.8.2
enum34==1.1.6
flake8==3.5.0
mccabe==0.6.1
pycodestyle==2.3.1
pyflakes==1.6.0
